@@ -7,7 +7,7 @@ const src = 'http://i.imgur.com/XhGsjTN.jpg';
 const imageStyle = {
   width: 800,
   margin: '0 auto'
-}
+};
 
 class App extends React.Component {
   render() {
@@ -15,17 +15,11 @@ class App extends React.Component {
       <div>
         <h1>Progressive!</h1>
         <ProgressiveImage src={src} placeholder={inline}>
-          {image => (
-            <img style={imageStyle} src={image}/>
-          )}
+          {image => <img style={imageStyle} src={image} />}
         </ProgressiveImage>
       </div>
-    )
+    );
   }
 }
 
-
-ReactDOM.render(
-  <App />,
-  document.getElementById('content')
-);
+ReactDOM.render(<App />, document.getElementById('content'));
